simpletextextractfromworddoc/demo/requirements.txt
gradio_simpletextextractfromworddoc